job_scraper_python/requirements.txt
PyYAML~=6.0
python-dotenv~=1.0.0
playwright~=1.40
pandas~=2.0
openpyxl~=3.1
PyYAML~=6.0
python-dotenv~=1.0.0
playwright~=1.40
pandas~=2.0
openpyxl~=3.1
requests~=2.28
